Structure Revamp with some changes
