+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 group = 'com.rosekingdom'
-version = '0.1.8'
+version = '0.2.0'
 
 repositories {
     mavenCentral()
@@ -15,10 +15,14 @@ repositories {
         name = "sonatype"
         url = "https://oss.sonatype.org/content/groups/public/"
     }
+    maven {
+        url "https://repo.dmulloy2.net/repository/public/"
+    }
 }
 
 dependencies {
     compileOnly "io.papermc.paper:paper-api:1.20.1-R0.1-SNAPSHOT"
+    compileOnly group: "com.comphenix.protocol", name: "ProtocolLib", version: "4.8.0"
 }
 
 def targetJavaVersion = 17
