+++ src/main/java/com/rosekingdom/rosekingdom/RoseKingdom.java
@@ -1,20 +1,20 @@
 package com.rosekingdom.rosekingdom;
 
-import com.rosekingdom.rosekingdom.Commands.Manager.CommandManager;
-import com.rosekingdom.rosekingdom.Config.Config;
-import com.rosekingdom.rosekingdom.Database.Database;
-import com.rosekingdom.rosekingdom.Events.EventHandler;
-import com.rosekingdom.rosekingdom.Premissions.Teams;
-import com.rosekingdom.rosekingdom.Utils.Grave;
-import com.rosekingdom.rosekingdom.Database.Statements.GraveStatement;
+import com.rosekingdom.rosekingdom.Core.CommandManager.CommandManager;
+import com.rosekingdom.rosekingdom.Core.Config.Config;
+import com.rosekingdom.rosekingdom.Core.Database.Database;
+import com.rosekingdom.rosekingdom.Core.Premissions.Teams;
+import com.rosekingdom.rosekingdom.Core.Utils.EventHandler;
+import com.rosekingdom.rosekingdom.Graves.Grave;
+import com.rosekingdom.rosekingdom.Graves.Statements.DeathStatement;
 import org.bukkit.plugin.java.JavaPlugin;
 
-import java.sql.SQLException;
 
 public final class RoseKingdom extends JavaPlugin {
     public static int players=0;
     @Override
     public void onEnable() {
+        EventHandler.setProtocolManager();
         getLogger().info("RoseKingdom Started Loading!");
         new CommandManager(this);
 
@@ -48,10 +48,10 @@ public final class RoseKingdom extends JavaPlugin {
         int total_graves_loaded = 0;
 
         getLogger().info("Loading graves...");
-        for(int id : GraveStatement.getGraveOwners()){
-            for(String graveId : GraveStatement.getGraves(id)){
+        for(int id : DeathStatement.getGraveOwners()){
+            for(String graveId : DeathStatement.getGraves(id)){
                 Grave grave = new Grave(id, graveId);
-                grave.timer(GraveStatement.getTime(id, graveId));
+                grave.timer(DeathStatement.getTime(id, graveId));
                 Grave.addGrave(grave);
                 total_graves_loaded++;
             }
